+++ src/main/java/ru/practicum/shareit/constant/Constants.java
@@ -1,4 +1,4 @@
-package ru.practicum.shareit.item;
+package ru.practicum.shareit.constant;
 
 import lombok.NoArgsConstructor;
