+++ src/main/java/ru/practicum/shareit/item/model/Item.java
@@ -1,18 +1,32 @@
 package ru.practicum.shareit.item.model;
 
 import lombok.*;
+import org.mapstruct.ap.internal.model.GeneratedType;
 import ru.practicum.shareit.request.ItemRequest;
+import ru.practicum.shareit.user.model.User;
+
+import javax.persistence.*;
 
 @Data
 @Builder
+@Entity
+@Table(name = "items")
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
 public class Item {
 
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     private int id;
     private String name; // краткое название
     private String description; // развёрнутое описание
     private Boolean available; // статус о том, доступна или нет вещь для аренды
+    @Column(name = "owner_id")
     private int owner; // владелец вещи
     // если вещь была создана по запросу другого пользователя,
     // то в этом поле будет храниться ссылка на соответствующий запрос
-    private ItemRequest request;
+    @Column(name = "request_id")
+    private int request;
 }
