+++ src/main/resources/schema.sql
@@ -1,27 +1,28 @@
 CREATE TABLE IF NOT EXISTS users
 (
     id    INTEGER GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY NOT NULL,
-    name  VARCHAR(255)                             NOT NULL,
-    email VARCHAR(255)                             NOT NULL UNIQUE
+    name  VARCHAR(255)                                         NOT NULL,
+    email VARCHAR(255)                                         NOT NULL UNIQUE
 );
 
 CREATE TABLE IF NOT EXISTS items
 (
     id          INTEGER GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY NOT NULL,
-    name        VARCHAR(255)                             NOT NULL,
+    name        VARCHAR(255)                                         NOT NULL,
     description VARCHAR(500),
-    owner       INTEGER REFERENCES users (id),
-    available   BOOLEAN                                  NOT NULL
+    owner_id    INTEGER REFERENCES users (id),
+    available   BOOLEAN                                              NOT NULL,
+    request_id  INTEGER
 );
 
 CREATE TABLE IF NOT EXISTS booking
 (
-    id     INTEGER GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY NOT NULL,
-    start_date   TIMESTAMP WITHOUT TIME ZONE,
-    end_date     TIMESTAMP WITHOUT TIME ZONE,
-    item_id   INTEGER REFERENCES items (id),
-    booker_id INTEGER REFERENCES users (id),
-    status VARCHAR(255)
+    id         INTEGER GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY NOT NULL,
+    start_date TIMESTAMP WITHOUT TIME ZONE,
+    end_date   TIMESTAMP WITHOUT TIME ZONE,
+    item_id    INTEGER REFERENCES items (id),
+    booker_id  INTEGER REFERENCES users (id),
+    status     VARCHAR(255)
 );
 
 CREATE TABLE IF NOT EXISTS comments
